+++ src/search/probfd/heuristics/pdbs/cegar/pucs_flaw_finder.cc
@@ -171,7 +171,7 @@ bool PUCSFlawFinder::expand(
             const auto succ_prob = path_probability * outcome.get_probability();
             State successor = registry.get_successor_state(state, outcome);
 
-            if (registry.size() > max_search_states) {
+            if (static_cast<int>(registry.size()) > max_search_states) {
                 return false;
             }
